fix TestingFarmProvisioner OS cleaning for RHEL-7/Centos-7

+++ atex/provisioner/testingfarm/cleaning.sh
@@ -9,6 +9,12 @@ set -e -x
 os_id=$(. /etc/os-release; echo "$ID")
 os_version=$(. /etc/os-release; echo "$VERSION_ID")
 
+if command -v dnf >/dev/null; then
+    pkg_tool="dnf -y --setopt=install_weak_deps=False"
+else
+    pkg_tool="yum -y"
+fi
+
 # ------------------------------------------------------------------------------
 
 # remove tmt-related commands
@@ -26,7 +32,8 @@ rm -rf /root/bin /root/.reserved-until
 
 # ------------------------------------------------------------------------------
 
-if [[ ! -e /run/ostree-booted ]]; then
+# RHEL-8+ only
+if [[ ! -e /run/ostree-booted ]] && command -v dnf >/dev/null; then
     # remove useless daemons to free up RAM a bit
     dnf remove -y rng-tools irqbalance
 
@@ -50,13 +57,6 @@ fi
 
 # ------------------------------------------------------------------------------
 
-# replace fedora mirrormanager-based repositories with primary/master ones,
-# which tend to be a lot more reliable
-# - this is to avoid checksum errors that very commonly pop up on mirrormanager
-#   on all mirrors (so trying different mirrors doesn't help and dnf eventually
-#   fails):
-#     Downloading successful, but checksum doesn't match. Calculated: 1abb62...
-#     Expected: a91641...
 function mkrepo {
     echo "[$1]"
     echo "name=$1"
@@ -68,7 +68,13 @@ function mkrepo {
         echo "$additional"
     done
 }
-# 8 is on vault/archive, 10 is currently broken
+# replace fedora mirrormanager-based repositories with primary/master ones,
+# which tend to be a lot more reliable
+# - this is to avoid checksum errors that very commonly pop up on mirrormanager
+#   on all mirrors (so trying different mirrors doesn't help and dnf eventually
+#   fails):
+#     Downloading successful, but checksum doesn't match. Calculated: 1abb62...
+#     Expected: a91641...
 if [[ $os_id == centos && ( $os_version == 9 || $os_version == 10 ) ]]; then
     case "$os_version" in
         9)  variants="BaseOS AppStream CRB HighAvailability NFV RT ResilientStorage" ;;
@@ -82,6 +88,16 @@ if [[ $os_id == centos && ( $os_version == 9 || $os_version == 10 ) ]]; then
         mkrepo "centos-master-$variant-debuginfo" "https://mirror.stream.centos.org/\$stream/$variant/\$basearch/debug/tree/" enabled=0
         echo
     done > /etc/yum.repos.d/centos-master.repo
+
+# ... except for CS7/CS8, which is on vault
+elif [[ $os_id == centos && $os_version -le 8 ]]; then
+    if grep -q 'mirror\.centos\.org' /etc/yum.repos.d/CentOS-*.repo; then
+        sed -i \
+            -e 's/^mirrorlist/#mirrorlist/' \
+            -e 's/^#baseurl/baseurl/' \
+            -e 's/mirror\.centos\.org/vault.centos.org/' \
+            /etc/yum.repos.d/CentOS-*.repo
+    fi
 fi
 
 # ------------------------------------------------------------------------------
@@ -134,5 +150,5 @@ echo -n > /etc/tmpfiles.d/restraint.conf
 # - this fixes issues with outdated packages, and overall isn't that big,
 #   typically 10-30 packages are upgraded
 if [[ $os_id == rhel || $os_id == centos ]]; then
-    dnf upgrade -y --skip-broken || true
+    $pkg_tool upgrade --skip-broken || true
 fi
